+++ src/schick/seg077.h
@@ -5,7 +5,7 @@ namespace M302de {
 extern struct struct_chest g_dng01_specialchests[9];
 
 /* 0x20 */
-signed short DNG01_handler(void);
+signed int DNG01_handler(void);
 /* 0x25 */
 void DNG01_chest01_loot(struct struct_chest*);
 /* 0x2a */
