+++ src/schick/seg080.h
@@ -17,9 +17,9 @@ void DNG04_chest02_loot(struct struct_chest*);
 /* 0x39 */
 void DNG04_chest02_open(struct struct_chest*);
 /* 0x3e */
-signed short DNG04_handler(void);
+signed int DNG04_handler(void);
 /* 0x43 */
-signed short DNG05_handler(void);
+signed int DNG05_handler(void);
 /* 0x48 */
 void DNG05_fight_intro(const signed int);
