+++ src/schick/seg081.h
@@ -9,7 +9,7 @@ void DNG06_chest00_loot(struct struct_chest*);
 /* 0x25 */
 void DNG06_chest01_loot(struct struct_chest*);
 /* 0x2a */
-signed short DNG06_handler(void);
+signed int DNG06_handler(void);
 /* 0x2f */
 void DNG06_fight_intro(const signed int);
 /* 0x34 */
